cache findings for files
Calculate a sha256 sum of file content, and use that as the filename
of a cache file to store any findings. On subsequent runs if the file
content is unmodified, the findings are just loaded from the cache.

On a 540 MB repository, saving the cache increases the scan time by
25%, from 12 seconds to 15 seconds. With a fully pre-populated cache,
however, the scan time reduces by 95% to just 0.6 milli seconds.

XXX: invalidate basd on pattern file changes, or on gitleaks binary
changes.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>